# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: market
    description: Simulated market intelligence
  - name: journal
    description: Shadow trade journal
  - name: risk
    description: Risk guardrails
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /market/snapshot:
    get:
      operationId: getMarketSnapshot
      tags: [market]
      summary: Get the simulated market snapshot
      description: Returns simulated five-minute candles, levels, indicators, and discipline signals. This endpoint never connects to a broker or places orders.
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 12
            default: MES
        - name: session
          in: query
          required: false
          schema:
            type: string
            enum: [premarket, regular]
            default: regular
      responses:
        "200":
          description: Simulated market snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketSnapshot"
        "400":
          description: Invalid market query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /futures/contracts:
    get:
      operationId: listFuturesContractSpecifications
      tags: [market]
      summary: List configurable simulated futures contracts
      description: Returns the configurable contract catalog used by the deterministic Shadow Mode feed. Values must be verified before any future provider integration.
      responses:
        "200":
          description: Futures contract specifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FuturesContractSpecification"
  /dashboard/overview:
    get:
      operationId: getDashboardOverview
      tags: [market]
      summary: Get dashboard overview
      description: Returns the current shadow session summary and recent journal activity.
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverview"
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List shadow journal entries
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Record a shadow trade
      description: Records a simulated trade review only. It does not create a broker order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Journal entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "400":
          description: Invalid journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /journal/{id}:
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a shadow journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Journal entry deleted
        "404":
          description: Journal entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /risk/settings:
    get:
      operationId: getRiskSettings
      tags: [risk]
      summary: Get risk settings
      responses:
        "200":
          description: Risk settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskSettings"
    patch:
      operationId: updateRiskSettings
      tags: [risk]
      summary: Update risk settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskSettingsUpdate"
      responses:
        "200":
          description: Updated risk settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskSettings"
        "400":
          description: Invalid risk settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Candle:
      type: object
      properties:
        time:
          type: string
        timestamp:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
        isComplete:
          type: boolean
        bid:
          type: number
        ask:
          type: number
        bidSize:
          type: number
        askSize:
          type: number
        contractSymbol:
          type: string
      required: [time, timestamp, openTime, closeTime, open, high, low, close, volume, isComplete, bid, ask, bidSize, askSize, contractSymbol]
    SessionHours:
      type: object
      properties:
        timeZone:
          type: string
        start:
          type: string
        end:
          type: string
      required: [timeZone, start, end]
    FuturesContractSpecification:
      type: object
      properties:
        rootSymbol:
          type: string
        fullContractSymbol:
          type: string
        exchange:
          type: string
        contractMonth:
          type: string
        tickSize:
          type: number
        dollarValuePerTick:
          type: number
        pointValue:
          type: number
        contractMultiplier:
          type: number
        regularSessionHours:
          $ref: "#/components/schemas/SessionHours"
        commissionPerContract:
          type: number
        exchangeAndRegulatoryFeesPerContract:
          type: number
        maximumSpreadTicks:
          type: number
        minimumLiquidity:
          type: number
        rolloverDate:
          type: string
        configurable:
          type: boolean
        verificationNote:
          type: string
      required: [rootSymbol, fullContractSymbol, exchange, contractMonth, tickSize, dollarValuePerTick, pointValue, contractMultiplier, regularSessionHours, commissionPerContract, exchangeAndRegulatoryFeesPerContract, maximumSpreadTicks, minimumLiquidity, rolloverDate, configurable, verificationNote]
    FuturesSessionCalendar:
      type: object
      properties:
        timeZone:
          type: string
        premarket:
          $ref: "#/components/schemas/SessionHours"
        regular:
          $ref: "#/components/schemas/SessionHours"
      required: [timeZone, premarket, regular]
    Signal:
      type: object
      properties:
        key:
          type: string
          enum: [orb, pullback, patience, volume]
        label:
          type: string
        status:
          type: string
          enum: [confirmed, watching, blocked]
        detail:
          type: string
      required: [key, label, status, detail]
    ReplayMetadata:
      type: object
      properties:
        cursor:
          type: string
        visibleCandleCount:
          type: number
        timeZone:
          type: string
        barIntervalMinutes:
          type: number
          enum: [5]
      required: [cursor, visibleCandleCount, timeZone, barIntervalMinutes]
    CriticalLevel:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        kind:
          type: string
      required: [name, price, kind]
    NtzState:
      type: object
      properties:
        status:
          type: string
          enum: [pending, inside, outside]
        complete:
          type: boolean
      required: [status, complete]
    TrendEvidence:
      type: object
      properties:
        direction:
          type: string
          enum: [bullish, bearish, neutral]
        score:
          type: number
        evidence:
          type: array
          items:
            type: string
        structure:
          type: string
      required: [direction, score, evidence, structure]
    RuleEvidence:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        detail:
          type: string
      required: [key, label, detail]
    StrategyDecision:
      type: object
      properties:
        state:
          type: string
          enum: ["NO TRADE", "WAITING", "SETUP FORMING", "SETUP QUALIFIED", "POSSIBLE REVERSAL", "RISK LOCKOUT"]
        explanation:
          type: string
        passedRules:
          type: array
          items:
            $ref: "#/components/schemas/RuleEvidence"
        failedRules:
          type: array
          items:
            $ref: "#/components/schemas/RuleEvidence"
      required: [state, explanation, passedRules, failedRules]
    RiskPlan:
      type: object
      properties:
        direction:
          type: string
          enum: [long, short]
        entry:
          type: ["number", "null"]
        thesisStop:
          type: ["number", "null"]
        catastropheStop:
          type: ["number", "null"]
        target:
          type: ["number", "null"]
        contracts:
          type: number
        dollarRisk:
          type: number
        allowed:
          type: boolean
        reasons:
          type: array
          items:
            type: string
      required: [direction, entry, thesisStop, catastropheStop, target, contracts, dollarRisk, allowed, reasons]
    LevelStoryEvent:
      type: object
      properties:
        time:
          type: string
        level:
          type: string
        interaction:
          type: string
        detail:
          type: string
      required: [time, level, interaction, detail]
    ReversalState:
      type: object
      properties:
        doji:
          type: boolean
        equivalentCandles:
          type: boolean
        warning:
          type: ["string", "null"]
      required: [doji, equivalentCandles, warning]
    MarketSnapshot:
      type: object
      properties:
        mode:
          type: string
          enum: ["SHADOW MODE — NO LIVE ORDERS"]
        symbol:
          type: string
        company:
          type: string
        contract:
          $ref: "#/components/schemas/FuturesContractSpecification"
        sessionCalendar:
          $ref: "#/components/schemas/FuturesSessionCalendar"
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        marketStatus:
          type: string
          enum: [premarket, open, closed]
        session:
          type: string
        updatedAt:
          type: string
        replay:
          $ref: "#/components/schemas/ReplayMetadata"
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        levels:
          type: object
          properties:
            premarketHigh: { type: number }
            premarketLow: { type: number }
            previousDayHigh: { type: number }
            previousDayLow: { type: number }
            previousDayClose: { type: number }
            dayBeforeYesterdayHigh: { type: ["number", "null"] }
            dayBeforeYesterdayLow: { type: ["number", "null"] }
            openingRangeHigh: { type: ["number", "null"] }
            openingRangeLow: { type: ["number", "null"] }
            ntzHigh: { type: ["number", "null"] }
            ntzLow: { type: ["number", "null"] }
            ntzWidth: { type: ["number", "null"] }
            vwap: { type: ["number", "null"] }
            critical:
              type: array
              items:
                $ref: "#/components/schemas/CriticalLevel"
          required: [premarketHigh, premarketLow, previousDayHigh, previousDayLow, previousDayClose, dayBeforeYesterdayHigh, dayBeforeYesterdayLow, openingRangeHigh, openingRangeLow, ntzHigh, ntzLow, ntzWidth, vwap, critical]
        ntz:
          $ref: "#/components/schemas/NtzState"
        indicators:
          type: object
          properties:
            rsi: { type: ["number", "null"] }
            ema200: { type: ["number", "null"] }
            emaSlope: { type: ["number", "null"] }
            vwap: { type: ["number", "null"] }
            fib236: { type: ["number", "null"] }
            fib382: { type: ["number", "null"] }
            fib5: { type: ["number", "null"] }
            fib618: { type: ["number", "null"] }
            fib786: { type: ["number", "null"] }
            volumeRatio: { type: ["number", "null"] }
          required: [rsi, ema200, emaSlope, vwap, fib236, fib382, fib5, fib618, fib786, volumeRatio]
        trend:
          $ref: "#/components/schemas/TrendEvidence"
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        decision:
          $ref: "#/components/schemas/StrategyDecision"
        riskPlan:
          $ref: "#/components/schemas/RiskPlan"
        levelStory:
          type: array
          items:
            $ref: "#/components/schemas/LevelStoryEvent"
        reversal:
          $ref: "#/components/schemas/ReversalState"
        assumptions:
          type: array
          items:
            type: string
      required: [mode, symbol, company, contract, sessionCalendar, price, change, changePercent, marketStatus, session, updatedAt, replay, candles, levels, ntz, indicators, trend, signals, decision, riskPlan, levelStory, reversal, assumptions]
    DashboardOverview:
      type: object
      properties:
        sessionPnl:
          type: number
        sessionPnlPercent:
          type: number
        maxDailyLoss:
          type: number
        dailyLossUsed:
          type: number
        tradeCount:
          type: number
        winRate:
          type: number
        checklistCompleted:
          type: number
        checklistTotal:
          type: number
        recentEntries:
          type: array
          items:
            $ref: "#/components/schemas/JournalEntry"
      required: [sessionPnl, sessionPnlPercent, maxDailyLoss, dailyLossUsed, tradeCount, winRate, checklistCompleted, checklistTotal, recentEntries]
    JournalEntry:
      type: object
      properties:
        id:
          type: number
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        setup:
          type: string
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        quantity:
          type: number
        pnl:
          type: ["number", "null"]
        notes:
          type: string
        checklistPassed:
          type: boolean
        createdAt:
          type: string
      required: [id, symbol, side, setup, entryPrice, exitPrice, quantity, pnl, notes, checklistPassed, createdAt]
    JournalEntryInput:
      type: object
      properties:
        symbol:
          type: string
          minLength: 1
          maxLength: 12
        side:
          type: string
          enum: [long, short]
        setup:
          type: string
          minLength: 1
          maxLength: 80
        entryPrice:
          type: number
          exclusiveMinimum: 0
        exitPrice:
          type: ["number", "null"]
          exclusiveMinimum: 0
        quantity:
          type: number
          minimum: 1
        pnl:
          type: ["number", "null"]
        notes:
          type: string
          maxLength: 2000
        checklistPassed:
          type: boolean
      required: [symbol, side, setup, entryPrice, quantity, notes, checklistPassed]
    RiskSettings:
      type: object
      properties:
        accountSize:
          type: number
        riskPercent:
          type: number
        maxDailyLoss:
          type: number
        dailyLossUsed:
          type: number
        isLocked:
          type: boolean
      required: [accountSize, riskPercent, maxDailyLoss, dailyLossUsed, isLocked]
    RiskSettingsUpdate:
      type: object
      properties:
        accountSize:
          type: number
          exclusiveMinimum: 0
        riskPercent:
          type: number
          exclusiveMinimum: 0
          maximum: 100
        maxDailyLoss:
          type: number
          exclusiveMinimum: 0
      minProperties: 1