# lib/api-zod/src/generated/types/candle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Candle {
  time: string;
  timestamp: string;
  openTime: string;
  closeTime: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
  isComplete: boolean;
  bid: number;
  ask: number;
  bidSize: number;
  askSize: number;
  contractSymbol: string;
}