# lib/api-zod/src/generated/types/futuresContractSpecification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionHours } from './sessionHours';

export interface FuturesContractSpecification {
  rootSymbol: string;
  fullContractSymbol: string;
  exchange: string;
  contractMonth: string;
  tickSize: number;
  dollarValuePerTick: number;
  pointValue: number;
  contractMultiplier: number;
  regularSessionHours: SessionHours;
  commissionPerContract: number;
  exchangeAndRegulatoryFeesPerContract: number;
  maximumSpreadTicks: number;
  minimumLiquidity: number;
  rolloverDate: string;
  configurable: boolean;
  verificationNote: string;
}