# lib/api-zod/src/generated/types/futuresSessionCalendar.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionHours } from './sessionHours';

export interface FuturesSessionCalendar {
  timeZone: string;
  premarket: SessionHours;
  regular: SessionHours;
}