# lib/api-zod/src/generated/types/marketSnapshot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Candle } from './candle';
import type { FuturesContractSpecification } from './futuresContractSpecification';
import type { FuturesSessionCalendar } from './futuresSessionCalendar';
import type { LevelStoryEvent } from './levelStoryEvent';
import type { MarketSnapshotIndicators } from './marketSnapshotIndicators';
import type { MarketSnapshotLevels } from './marketSnapshotLevels';
import type { MarketSnapshotMarketStatus } from './marketSnapshotMarketStatus';
import type { MarketSnapshotMode } from './marketSnapshotMode';
import type { NtzState } from './ntzState';
import type { ReplayMetadata } from './replayMetadata';
import type { ReversalState } from './reversalState';
import type { RiskPlan } from './riskPlan';
import type { Signal } from './signal';
import type { StrategyDecision } from './strategyDecision';
import type { TrendEvidence } from './trendEvidence';

export interface MarketSnapshot {
  mode: MarketSnapshotMode;
  symbol: string;
  company: string;
  contract: FuturesContractSpecification;
  sessionCalendar: FuturesSessionCalendar;
  price: number;
  change: number;
  changePercent: number;
  marketStatus: MarketSnapshotMarketStatus;
  session: string;
  updatedAt: string;
  replay: ReplayMetadata;
  candles: Candle[];
  levels: MarketSnapshotLevels;
  ntz: NtzState;
  indicators: MarketSnapshotIndicators;
  trend: TrendEvidence;
  signals: Signal[];
  decision: StrategyDecision;
  riskPlan: RiskPlan;
  levelStory: LevelStoryEvent[];
  reversal: ReversalState;
  assumptions: string[];
}