# lib/api-zod/src/generated/types/replayMetadata.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReplayMetadataBarIntervalMinutes } from './replayMetadataBarIntervalMinutes';

export interface ReplayMetadata {
  cursor: string;
  visibleCandleCount: number;
  timeZone: string;
  barIntervalMinutes: ReplayMetadataBarIntervalMinutes;
}