# lib/api-zod/src/generated/types/riskPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskPlanDirection } from './riskPlanDirection';

export interface RiskPlan {
  direction: RiskPlanDirection;
  /** @nullable */
  entry: number | null;
  /** @nullable */
  thesisStop: number | null;
  /** @nullable */
  catastropheStop: number | null;
  /** @nullable */
  target: number | null;
  contracts: number;
  dollarRisk: number;
  allowed: boolean;
  reasons: string[];
}